Other Services

+++ src/main/java/tn/esprit/khairislimanitest/repositories/FactureRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.khairislimanitest.entities.Facture;
 
 
-public interface FactureRepository extends JpaRepository<Facture, Integer> {
+public interface FactureRepository extends JpaRepository<Facture, Long> {
 }
