+++ src/main/java/tn/esprit/khairislimanitest/repositories/FournisseurRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.khairislimanitest.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.khairislimanitest.entities.Fournisseur;
 
-public interface FournisseurRepository extends JpaRepository<Fournisseur, Integer> {
+public interface FournisseurRepository extends JpaRepository<Fournisseur, Long> {
 }
