+++ src/main/java/tn/esprit/khairislimanitest/repositories/RayonRepository.java
@@ -3,6 +3,6 @@ package tn.esprit.khairislimanitest.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.khairislimanitest.entities.Rayon;
 
-public interface RayonRepository extends JpaRepository<Rayon, Integer> {
+public interface RayonRepository extends JpaRepository<Rayon, Long> {
 
 }
