+++ src/main/java/tn/esprit/khairislimanitest/repositories/StockRepository.java
@@ -3,5 +3,6 @@ package tn.esprit.khairislimanitest.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.khairislimanitest.entities.Stock;
 
-public interface StockRepository extends JpaRepository<Stock, Integer> {
+public interface StockRepository extends JpaRepository<Stock, Long> {
+
 }
